+++ server/http_status.go
@@ -36,7 +36,6 @@ import (
 	"github.com/pingcap/parser/terror"
 	"github.com/pingcap/tidb/config"
 	"github.com/pingcap/tidb/kv"
-	"github.com/pingcap/tidb/server/rpcserver"
 	"github.com/pingcap/tidb/util"
 	"github.com/pingcap/tidb/util/logutil"
 	"github.com/pingcap/tidb/util/printer"
@@ -281,7 +280,7 @@ func (s *Server) setupStatuServerAndRPCServer(addr string, serverMux *http.Serve
 	grpcL := m.Match(cmux.Any())
 
 	s.statusServer = &http.Server{Addr: addr, Handler: CorsHandler{handler: serverMux, cfg: s.cfg}}
-	s.grpcServer = rpcserver.CreateTiDBRPCServer(s.cfg.Security)
+	s.grpcServer = CreateTiDBRPCServer(s.cfg.Security)
 
 	go util.WithRecovery(func() {
 		err := s.grpcServer.Serve(grpcL)
