+++ server/rpc_server.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package rpcserver
+package server
 
 import (
 	"github.com/crazycs520/diagnostics"
